+++ src/solvers/enumpoly/nfgpoly.cc
@@ -120,7 +120,7 @@ EnumPolyStrategySupportSolve(const StrategySupportProfile &support, bool &is_sin
   std::list<Vector<double>> roots;
   try {
     roots =
-        solver.FindRoots({bottoms, tops}, p_stopAfter.value_or(std::numeric_limits<int>::max()),
+        solver.FindRoots({bottoms, tops}, p_stopAfter.value_or(std::numeric_limits<size_t>::max()),
                          p_maxRectangles, p_budgetExceeded, p_cancel);
   }
   catch (const SingularMatrixException &) {
